CURA-5035 Renaming
