+++ plugins/Toolbox/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2018 Ultimaker B.V.
 # Toolbox is released under the terms of the LGPLv3 or higher.
 
-from .Toolbox import Toolbox
+from .src import Toolbox
 
 
 def getMetaData():
