+++ plugins/Toolbox/resources/qml/ToolboxInstalledTile.qml
