+++ plugins/Toolbox/resources/qml/ToolboxShadow.qml
